Extract search to service object
